git-svn-id: http://svn.perdix.co/svn/perdix/irf-perdix-client/branches/v8.1-KGFS_Production@7406 15cf4ce8-1186-411e-815e-a2b7fc64482c

+++ source/dev-www/modules/irfpages/states/Page.EngineOffline.js
@@ -7,6 +7,7 @@ irf.pages.controller("PageEngineOfflineCtrl", [
 	"irfStorageService",
 	"elementsUtils",
 	"entityManager",
+	"Utils",
 function(
 	$log,
 	$scope,
@@ -15,7 +16,8 @@ function(
 	$injector,
 	irfStorageService,
 	elementsUtils,
-	entityManager
+	entityManager,
+	Utils
 ) {
 	var self = this;
 	$log.info("Page.EngineOffline.html loaded");
@@ -33,12 +35,6 @@ function(
 		updateAppTitle($scope.page.title);
 	};
 
-	$scope.callback = function(item, index) {
-		$log.debug("Restoring offline data");
-		entityManager.setModel($scope.pageName, item);
-		$scope.loadPage(null);
-	};
-
 	if ($scope.page.offline && angular.isFunction($scope.page.getOfflineDisplayItem)) {
 		var items = irfStorageService.getMasterJSON($scope.pageName);
 
@@ -46,7 +42,7 @@ function(
 		var idx = 0;
 		_.forEach(items, function(value, key) {
 			offlineItems[idx] = value;
-			displayItems[idx] = $scope.page.getOfflineDisplayItem(value, idx);
+			try {displayItems[idx] = $scope.page.getOfflineDisplayItem(value, idx);} catch (e) {displayItems[idx] = ['PARSING_ERROR', e.message];}
 			for (var i = 0; i < displayItems[idx].length; i++) {
 				if (angular.isNumber(displayItems[idx][i]))
 					displayItems[idx][i] = displayItems[idx][i].toString();
@@ -63,12 +59,25 @@ function(
 
 	$scope.offlineListInfo = {
 		actions: [{
+			name: "Open",
+			icon: "fa fa-folder-open-o",
+			fn: function(item, index){
+				$log.debug("Restoring offline data");
+				entityManager.setModel($scope.pageName, item);
+				$scope.loadPage(null);
+			},
+			isApplicable: function(item, index){
+				return true;
+			}
+		}, {
 			name: "Delete",
 			icon: "fa fa-trash",
 			fn: function(item, index){
-				_.pullAt($scope.displayItems, index);
-				_.pullAt($scope.offlineItems, index);
-				irfStorageService.deleteJSON($scope.pageName, item.$$STORAGE_KEY$$);
+				Utils.confirm("Are You Sure?").then(function(){
+					_.pullAt($scope.displayItems, index);
+					_.pullAt($scope.offlineItems, index);
+					irfStorageService.deleteJSON($scope.pageName, item.$$STORAGE_KEY$$);
+				});
 			},
 			isApplicable: function(item, index){
 				//if (index%2==0){
